Create CPU scaling component stub
Creates a stub implementation for CPU scaling component connected to
CPUScalingConfiguration API.
CPUScalingManager is the overarching interface for controlling workers
on assigned CPUs and accepting new scaling configurations.
CPUScalingWorker is run asynchronously and is responsible
for storing the scaling configuration tied to the CPU and for
calling the updater.
CPUScalingUpdater is run as a part of worker and is responsible for
performing actual scaling tasks based on scaling configuration
received from the worker.