Use hardware frequency limits when not specified in PowerProfile
This makes it possible to use hardware frequency limits in case
max and min frequency values are not specified in the PowerProfile.

This was previously only supported for PowerProfiles that specify
the epp value, which were assumed to be base PowerProfiles, though
it worked for user-defined PowerProfiles as well, as long as they
specified the epp. This restriction is now removed.

Max and min frequency values must be both specified or omitted. If only
one of these values is specified, a validation error will be returned
by the reconciler.

Signed-off-by: Aleksander Polak <aleksander.polak@amd.com>