Handle scheme properly in tests
This prevents tests from mutating the global scheme object which caused
some of them to only work when executed in a specific order.

A new scheme object is now created for each test at setup time.
Required APIs are registered to this scheme.

Additionally, this unifies and cleans up the surrounding code across
all tests.

Signed-off-by: Aleksander Polak <aleksander.polak@amd.com>