Bypass calls to E-SMI in controller tests
This is a workaround which allows bypassing calls to E-SMI
in `initUncore()` in all controller tests. This was achieved
by introducing a dedicated flag and implementing `TestMain`[^1]
for the controller package which sets this flag as part of the global
setup.

This is by no means a final solution, but it's cheap and more reliable
than the workaround we've used so far, which didn't truly isolate
E-SMI in tests.

[^1]: https://pkg.go.dev/testing#hdr-Main

Signed-off-by: Aleksander Polak <aleksander.polak@amd.com>