Support percentage values in PowerProfile (#66)
* Support percentage values in PowerProfile

This allows specifying values for the `.spec.max` and `.spec.min` fields
in PowerProfile as percentage, for example:

```yaml
apiVersion: "power.intel.com/v1"
kind: PowerProfile
metadata:
  name: my-profile
  namespace: power-manager
spec:
  name: my-profile
  max: 75%
  min: 25%
```

The allowed values are between 0% and 100% and indicate the percentage
of the supported hardware frequency range. The general formula is:

f = ceil(f_min + p (f_max - f_min))

In the formula above, f is the calculated frequency, f_min
is the minimum supported hardware frequency, f_max is the maximum
supported hardware frequency, and p is the percentage value. The result
is rounded up to the nearest integer.

For example, if the range is between 400 MHz and 3709 MHz inclusive,
the values 0%, 50%, and 100% correspond to 400 MHz, 2055 MHz,
and 3709 MHz, respectively.

---------

Signed-off-by: Aleksander Polak <aleksander.polak@amd.com>
Signed-off-by: Rafal Lal <rafallal@amd.com>
Co-authored-by: Rafal Lal <rafallal@amd.com>