Add fallback to scaling_max_freq without userspace governor
On platforms where the userspace governor is not supported,
i.e. AMD P-State EPP, attempts to set CPU frequency via
`scaling_setspeed` fail.

This patch enhances `setCPUFrequency()` to detect when the
userspace governor is not active and gracefully fall back to
using `scaling_max_freq` for setting the desired frequency.

Signed-off-by: Sivaprasad Tummala <sivaprasad.tummala@amd.com>